Extract scoped_bson_value into its own component
